# MobileApp/SeenIT/SeenIT/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RestApiSeenit",
    "version": "1.0"
  },
  "paths": {
    "/api/Film": {
      "get": {
        "tags": [
          "Film"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilmForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilmForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilmForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Film"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FilmForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FilmForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Film/{id}": {
      "get": {
        "tags": [
          "Film"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FilmForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilmForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Film"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FilmForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FilmForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Film"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FilmForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tytul": {
            "type": "string",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "dataPremiery": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "rodzajId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rodzajData": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}